decouple loading from HEPData from ID store
