+++ src/stores/storeid.ts
@@ -1,7 +1,5 @@
 import { defineStore } from 'pinia';
 import { useWorkspaceStore } from 'src/stores/workspace';
-import type { IAnalysis, IHEPdataentry } from 'src/interfaces';
-import { Notify } from 'quasar';
 
 export const useStoreIDStore = defineStore('storeids', {
   state: () => ({
@@ -21,62 +19,6 @@ export const useStoreIDStore = defineStore('storeids', {
       const id = this.add_store_with_id();
       const workspace_store = useWorkspaceStore(id)();
       workspace_store.load_workspace_from_url(url, name);
-    },
-    async check_workspaces_on_HEPdata(
-      hepdata_id: string
-    ): Promise<IAnalysis[]> {
-      this.checking = true;
-      const hepdata_url =
-        'https://www.hepdata.net/record/ins' +
-        hepdata_id +
-        '?format=json&light=true';
-      let hepdata_entry: IHEPdataentry | null = null;
-      try {
-        hepdata_entry = await (await fetch(hepdata_url)).json();
-      } catch {
-        Notify.create({
-          message:
-            'Invalid JSON encountered. Are you sure you provided the correct HEPdata ID?',
-          color: 'negative',
-          icon: 'report_problem',
-          position: 'top',
-        });
-        return [];
-      }
-      const analyses: IAnalysis[] | undefined = [];
-      if (hepdata_entry?.record.analyses === undefined) {
-        return [];
-      }
-      let analysis_index = -1;
-      for (const analysis of hepdata_entry?.record.analyses) {
-        analysis_index++;
-        if (analysis.type !== 'HistFactory') {
-          continue;
-        }
-        analyses.push({
-          name: hepdata_entry.resources_with_doi[analysis_index].filename,
-          url: analysis.analysis,
-        });
-      }
-      if (analyses === undefined || analyses.length === 0) {
-        Notify.create({
-          message:
-            'No JSON workspaces are available for this HEPData entry. Are you sure you provided the correct HEPdata ID?',
-          color: 'negative',
-          icon: 'report_problem',
-          position: 'top',
-        });
-        return [];
-      }
-      this.checking = false;
-      return analyses;
-    },
-    async load_workspaces_from_HEPdata(analyses: IAnalysis[]): Promise<void> {
-      for (const analysis of analyses) {
-        const id = this.add_store_with_id();
-        const workspace_store = useWorkspaceStore(id)();
-        workspace_store.load_workspace_from_HEPdata(analysis);
-      }
     },
     remove_store_with_id(id: number): void {
       const index = this.ids.indexOf(id);
